zig json-ast: introduce ast structs
